Text Sentiment Analyzer
This project was for my "Programming with Data" module, and is intended to scrape text from the web and conduct a sentiment analysis on it. The text scraped was presidential speeches, and the goal here is to compare the sentiments of the speeches with the overall "Happiness Index" of the related country.